# Restaurante/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Restaurante",
    "description": "Clase: IS-513 Seccion 1700  Proyecto Grupo Restaurante. Este API maneja las operaciones de un restaurante._\n\nLinks:\n- [Repositorio del proyecto](https://github.com/AdrianJMendez/Restaurante)",
    "contact": {
      "email": "viktor.hernandez@unah.hn"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8091"
    }
  ],
...
  "paths": {
...
    "/api/factura/obtener/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "summary": "Buscar la lista de facturas por cliente",
        "description": "Devuelve una lista",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "description": "Id del cliente por el que se va a buscar la factura",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Factura"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/factura/obtener/{facturaId}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "summary": "Buscar la factura por su id",
        "description": "Devuelve una lista",
        "parameters": [
          {
            "name": "facturaId",
            "in": "path",
            "description": "Id de la factura a obtener",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Factura"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/factura/obtener/cliente/nombre/{nombreCliente}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "summary": "Buscar la factura por su un nombre de Cliente",
        "description": "Devuelve una lista",
        "parameters": [
          {
            "name": "nombreCliente",
            "in": "path",
            "description": "nombre del cliente",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Factura"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/factura/crear": {
      "post": {
        "tags": [
          "Factura"
        ],
        "summary": "Crear una factura",
        "description": "Crear una factura",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuarioId": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Id del usuario que genera la factura",
                    "example": 1
                  },
                  "clienteId": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Id del cliente que va a pagar",
                    "example": 1
                  },
                  "fecha": {
                    "type": "string",
                    "description": "Fecha y hora de la emision de la factura",
                    "example": "2021-12-31 18:30"
                  },
                  "metodoDePago": {
                    "type": "string",
                    "description": "Metodo con el que el cliente efectuo el pago",
                    "example": "efectivo"
                  },
                  "platillos": {
                    "type": "array",
                    "description": "Platillos que lleva la orden facturada",
                    "items": {
                      "oneOf": [
                        {
                          "type": "object",
                          "properties": {
                            "platilloId": {
                              "type": "string",
                              "example": 1
                            }
                          }
                        },
                        {
                          "type": "integer",
                          "example": "cantidad=2"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Factura"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/factura/borrar/{facturaId}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "summary": "Borra la factura por su id",
        "description": "Borra una factura",
        "parameters": [
          {
            "name": "facturaId",
            "in": "path",
            "description": "Id de la factura que se va a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Factura eliminada"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
...
    "/api/restaurante/cliente/obtener/identifiacion/{identificacion}": {
      "get": {
        "tags": [
          "Cliente"
        ],
        "summary": "Buscar el cliente por su identificion",
        "description": "Devuelve un cliente",
        "parameters": [
          {
            "name": "identificacion",
            "in": "path",
            "description": "RTN del cliente",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/cliente/crear": {
      "post": {
        "tags": [
          "Cliente"
        ],
        "summary": "Crear un cliente",
        "description": "Crear un cliente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del cliente",
                    "example": "Jose"
                  },
                  "apellido": {
                    "type": "string",
                    "description": "Apellido del cliente",
                    "example": "Martinez"
                  },
                  "identificacion": {
                    "type": "string",
                    "description": "RTN del cliente",
                    "example": 801198309594
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/cliente/borrar/id/{clienteId}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "summary": "Borra el cliente por su id",
        "description": "Borra un cliente",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "description": "Id del cliente a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Cliente eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/cliente/borrar/dni/{dni}": {
      "delete": {
        "tags": [
          "Cliente"
        ],
        "summary": "Borra el inventario por su nombre",
        "description": "Borra un inventario",
        "parameters": [
          {
            "name": "dni",
            "in": "path",
            "description": "DNI del cliente a eliminar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion Exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Respuesta": {
                      "type": "string",
                      "example": "Cliente eliminado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/cliente/actualizar/id/{clienteId}": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "summary": "Actulizar un cliente",
        "description": "Actualizar un cliente",
        "parameters": [
          {
            "name": "clienteId",
            "in": "path",
            "description": "Id del inventario a actualizar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del cliente",
                    "example": "Miguel"
                  },
                  "apellido": {
                    "type": "string",
                    "description": "Apellido del cliente",
                    "example": "Laitano"
                  },
                  "identificacion": {
                    "type": "string",
                    "description": "DNI del cliente",
                    "example": "0715198502912"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/restaurante/cliente/actualizar/dni/{dni}": {
      "put": {
        "tags": [
          "Cliente"
        ],
        "summary": "Actulizar un cliente",
        "description": "Actualizar un cliente",
        "parameters": [
          {
            "name": "dni",
            "in": "path",
            "description": "Nombre del cliente a actualizar",
            "required": true,
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string",
                    "description": "Nombre del cliente",
                    "example": "Miguel"
                  },
                  "apellido": {
                    "type": "string",
                    "description": "Apellido del cliente",
                    "example": "Laitano"
                  },
                  "identificacion": {
                    "type": "string",
                    "description": "DNI del cliente",
                    "example": "0715198502912"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cliente"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "clienteId": {
            "type": "integer",
            "description": "Generado por la base de datos",
            "format": "int64",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Primer nombre del cliente",
            "example": "Viktor"
          },
          "apellido": {
            "type": "string",
            "description": "Primer apellido del cliente",
            "example": "Hernandez"
          },
          "identificacion": {
            "type": "string",
            "description": "RTN del cliente",
            "example": 715200500005
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "usuarioId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "user": {
            "type": "string",
            "description": "Nombre del usuario",
            "example": "Adrian"
          },
          "categorioPermiso": {
            "$ref": "#/components/schemas/CategoriaPermiso"
          }
        }
      },
      "CategoriaPermiso": {
        "type": "object",
        "properties": {
          "categoriaPermisoId": {
            "type": "integer",
            "format": "int64",
            "description": "Solo hay 3 tipos de categorias",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la categoria",
            "example": "Gerente"
          },
          "descripcion": {
            "type": "string",
            "example": "Puede crear nuevos usuarios"
          }
        }
      },
      "Inventario": {
        "type": "object",
        "properties": {
          "inventarioId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del item del inventario",
            "example": "arroz"
          },
          "precioCompra": {
            "type": "number",
            "description": "Precio de compra del item por unidad",
            "example": 10
          },
          "cantidad": {
            "type": "number",
            "description": "Cantidad del item que se encuentra en existencia",
            "example": 10
          },
          "unidadMetrica": {
            "type": "string",
            "description": "Unidad con la que se mide el item",
            "example": "libra"
          },
          "minimoRecompra": {
            "type": "number",
            "description": "Cantidad minima del item que debe haber en existencia",
            "example": 5
          }
        }
      },
      "Platillo": {
        "type": "object",
        "properties": {
          "platilloId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del platillo",
            "example": "Bistec"
          },
          "precio": {
            "type": "number",
            "description": "Precio de venta del platillo",
            "example": 100
          },
          "imagen": {
            "type": "string",
            "description": "URL de la imagen a mostrar en el frontend del platillo",
            "example": "bisctec.jpg"
          },
          "inventarios": {
            "type": "array",
            "description": "Items del inventario que lleva la receta del platillo",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Inventario"
                },
                {
                  "type": "integer",
                  "description": "Cantidad del item del inventario que lleva el platillo",
                  "example": "cantidad = 2"
                }
              ]
            }
          }
        }
      },
      "Factura": {
        "type": "object",
        "properties": {
          "facturaId": {
            "type": "integer",
            "format": "int64",
            "description": "Generado por la base de datos",
            "example": 1
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "fecha": {
            "type": "string",
            "description": "Fecha y hora de la emision de la factura",
            "example": "2021-12-31 18:30"
          },
          "metodoDePago": {
            "type": "string",
            "description": "Metodo con el que el cliente efectuo el pago",
            "example": "efectivo"
          },
          "impuesto": {
            "type": "number",
            "description": "Impuesto (25% del subtotal) Generado por la base de datos",
            "example": 150
          },
          "total": {
            "type": "number",
            "description": "Total a pagar(incluye impuesto) Generado por la base de datos",
            "example": 750
          },
          "platillos": {
            "type": "array",
            "description": "Platillos que lleva la orden facturada",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Platillo"
                },
                {
                  "type": "integer",
                  "description": "Cantidad del item del inventario que lleva el platillo",
                  "example": "cantidad = 2"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}